requirements.txt
asgiref==3.4.1
Django==3.2.9
django-crontab==0.7.1
django-rest-auth==0.9.5
djangorestframework==3.12.4
pytz==2021.3
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
uritemplate==4.1.1
